Absorb error client in error comparison client
